Integrated spring social but facing hibernate problem

+++ src/main/java/com/anizzzz/product/sssweaterhouse/repository/ProductSizeRepository.java
@@ -3,6 +3,6 @@ package com.anizzzz.product.sssweaterhouse.repository;
 import com.anizzzz.product.sssweaterhouse.model.ProductSize;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductSizeRepository extends JpaRepository<ProductSize,Integer> {
+public interface ProductSizeRepository extends JpaRepository<ProductSize,Long> {
     ProductSize findBySize(String size);
 }
